Add operation dependency selector check
Before this commit, an operation such as

{
    description: "foo",
    selectors: [],
    action: (bar) => {
        bar.textContent = "baz";
    }
}

would compile, but would crash with a cryptic error because bar is
undefined in the action body (because the action is always called with
the same number of arguments as the number of selectors).

This commit adds a check which ensures that an action cannot have an
arity greater than the number of dependency selectors accompanying it.
The code above still compiles, but it is guaranteed to crash
immediately, every time, with a much more descriptive error message.