Add USERSCRIPTER_VERBOSE hint
If any Userscripter-specific error arises, it might be useful for the
user to enable the USERSCRIPTER_VERBOSE flag to understand the error, so
we now produce a hint suggesting that.

Syntax error, type errors etc in the business code don't trigger the
hint, because enabling the verbose flag wouldn't help anyway.