Remove webpack-userscript dependency
After weeks of debugging Better SweClockers' dark theme feature in
collaboration with users, we realized that the problem only occurred
with Tampermonkey, not with Violentmonkey.

https://github.com/SimonAlling/better-sweclockers/issues/62

However, turns out it was our fault, not Tampermonkey's: We were
emitting metadata containing

    // @run_at document-start

instead of

    // @run-at document-start

because I forgot that we used to rely on the userscript-metadata package
to replace underscores with hyphens.