feat: Make date suffix in version optional (#38)
The `dateAsSemver` suffix automatically appended to the version when building in development mode can cause trouble when using Userscripter to create a WebExtension. For example, Chrome refuses to load a WebExtension with a version not formatted like `x.x.x`. One could build using `USERSCRIPTER_MODE=production` as a workaround, but that also minifies the code, making debugging/E2E testing harder.

Co-authored-by: Simon Alling <alling.simon@gmail.com>