refactor: Replace node-sass with sass (#45)
[Node Sass has been deprecated in favor of Dart Sass.][libsass]

For this project in particular, Node Sass is the cause of lengthy
compilations when running `npm ci`, annoying GYP errors and [multiple
security vulnerabilities][snyk].

[libsass]: https://sass-lang.com/blog/libsass-is-deprecated
[snyk]: https://snyk.io/test/npm/userscripter/1.4.0

When we first tried to migrate in #29, it turned out that ["just
[replacing] `node-sass` in [our] `package.json` file with
`sass`"][migrate] didn't work, because doing so broke our mechanism for
accessing JavaScript values from within SCSS (i.e. `getGlobal`). For
example, tests would fail like so:

    can expose data to SASS

    expect(received).toMatchInlineSnapshot(snapshot)

    Snapshot name: `can expose data to SASS 1`

    Snapshot: SassNumber {}
    Received: undefined

[migrate]: https://sass-lang.com/blog/libsass-is-deprecated#how-do-i-migrate

We tried to fix this by upgrading packages (e.g. `node-sass-utils` to
version 1.1.3 and `sass-loader` to 11.0.1), but that would cause build
errors in dependent userscripts like [Better SweClockers][bsc] (e.g.
`SassError: Only 0 arguments allowed, but 1 was passed.` at `getGlobal`
call sites). As far as we could tell, we wouldn't be able to fix that by
upgrading dependencies, at least not without forcing dependent projects
to upgrade to webpack 5.

[bsc]: https://github.com/simonalling/better-sweclockers

The solution we finally arrived at was to adapt `getGlobalFrom` to work
with `sass` instead of `node-sass` and, in the webpack config we
generate for dependent projects, make sure the parameters of the SASS
variable getter are properly Dart Sass-encoded – conceptually:

```diff
 sassOptions: {
     functions: {
-        ["getGlobal"]: getGlobal,
+        ["getGlobal($x0)"]: getGlobal,
     },
 },
```

We decided that this is _not_ a breaking change, because although
dependent projects using the SASS variable getter in JavaScript (i.e.
not in SCSS) may break, we don't consider that a supported use case (and
we don't know of any such projects). Dependent projects using
`getGlobal` as intended (i.e. in SCSS) shouldn't be affected. In order
to confirm this, we have built, tested and used Better SweClockers (as
of [`f01d834`][bsc-commit]) based on this PR without any issues.

[bsc-commit]: https://github.com/SimonAlling/better-sweclockers/commit/f01d8343e0f6b81a4a987c620c2a1ed6c1499edf

Co-authored-by: Simon Alling <alling.simon@gmail.com>