build: replace awesome-typescript-loader with ts-loader
awesome-typescript-loader doesn't work with modern TypeScript versions
etc and generally has become a massive pain point as soon as I try to do
just about anything. It hasn't seen a new release since 2018 and the
GitHub repo was archived on 2019-12-01, so I think it's safe to think of
it as abandoned.

ts-loader 8.4.0 is the latest version that works with webpack 4.

💡 `git show --color-words='[a-z-]+|.'`