fix: Add lockfile to bootstrapped userscript (#123)
A problem with the bootstrapped userscript is that the installation of its dependencies isn't deterministic, since it doesn't have a `package-lock.json`. This PR adds one, which in turn enables replacing `npm install` with `npm ci` in CI and documentation. I created the lockfile by running `npm install` in `bootstrap/` with npm 8.19.2 on Node 16.18.0.

I was unable to make `npm pack` include the lockfile if its name was `package-lock.json`, so I used the same workaround as in #67, namely using a different name for it and renaming it at run time.

This should be a backward compatible change, since `npm install` should still work.

## About the diff

The `name` field is added in `bootstrap/package.json` because otherwise `npm install` sets it to `"bootstrap"` (or whatever directory the bootstrapped userscript lives in) in `bootstrap/package-lock.json`.

💡 `git show --color-words`