docs: Document release workflow (#126)
I've used this workflow for the last couple of releases. It's intentionally designed in a merge-then-publish fashion, so that we never end up publishing a package that wasn't built from a commit on the trunk. The downside is that it's possible to merge and tag a release commit from which we cannot publish, but I think that's much more acceptable.