feat!: Remove CLI tool; use tiged to bootstrap (#131)
Today, Userscripter can be said to consist of two parts:

  * a library, in turn consisting of a build-time part (`src/build/`) and a run-time part (`src/lib/`); and
  * a CLI tool, used to create new userscripts (`src/bin/`).

An existing userscript only needs the library. The CLI tool is annoyingly complex and adds dependencies and code that are completely unnecessary for users of the library.

This PR removes the CLI tool and recommends `tiged` (a `degit` fork) instead of `userscripter init` for bootstrapping a new userscript, bringing these benefits:

  * No CLI tool to maintain.
  * Fewer dependencies.
  * npm package need not include `bootstrap/`.
  * Smaller package size (134 kB → 23 kB compressed; 493 kB → 88 kB unpacked).
  * `bootstrap/.gitignore` and `bootstrap/package-lock.json` can simply have those names instead of being renamed at `userscripter init`-time because they don't survive into the bootstrapped userscript otherwise.

`userscripter init` uses the content of `bootstrap/` as it looked when the user's _installed_ version of the `userscripter` package was _released_. The `tiged` command in `README.md` instead uses `bootstrap/` as it looks on `master` whenever the command is _run_. In practice, both the old and the new recommended way to bootstrap a userscript should typically use the most up-to-date content of `bootstrap/`.

Just like today, the bootstrapped userscript will use whatever version of the library that `bootstrap/package.json` specifies, which may be outdated. Fixing that is for a future PR.

💡 `git show --color-words='userscripter init|.' -- ':!package-lock.json'`