ci: Make bootstrapped userscript dependent (#138)
The `bootstrap` job doesn't check how any changes to the library/package might affect the bootstrapped userscript; it only checks changes to the bootstrapped userscript itself (i.e. in `bootstrap/`). But we do want to know if a PR breaks the bootstrapped userscript, because then the PR may need to be considered breaking.

This PR therefore makes the bootstrapped userscript a so-called dependent userscript, just like Better SweClockers and Example Userscript. Building dependent userscripts deliberately isn't a _gating_ check because, as previously explained in #68, the purpose isn't to prevent _deliberate_ breaking changes, just _accidental_ ones.

💡 `git show --color-words=.`