docs: Update bootstrapped userscript post-release (#141)
I used this workflow in #140. It ensures that the recommended way of bootstrapping a new userscript (the `tiged` command in `README.md`) will always work, including when a release PR has been merged but a corresponding version of the package hasn't been published to the npm registry (either because it hasn't been done yet or because it cannot be done at all for whatever reason).