build: Upgrade to TypeScript 4.4 (#146)
This will enable upgrading to Jest 26 (#144; see #145).

## Breaking?

See #145 for details on why this should possibly be considered a breaking change.

I consider it backward compatible, because I don't think any consumer is broken by it. Notably, building with and without this PR in different Git working trees and using `git diff --no-index` to diff the emitted code (`build/` and `lib/`) reveals that the only difference is that some expressions in `.js` files are transformed from `foo` to `(0, foo)`.